moved styles to differnt files. -style adjustments for charts
